% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.r
\name{accumulate}
\alias{accumulate}
\title{Accumulate intermediate results of a vector reduction}
\usage{
accumulate(.x, .f, ..., .init)
}
\description{
Accumulate intermediate results of a vector reduction
}
