% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R, R/as_mapper.R
\name{as_mapper}
\alias{as_mapper}
\title{Convert an object into a mapper function}
\usage{
as_mapper(.f, ...)

as_mapper(.f, ...)
}
\description{
Convert an object into a mapper function

Convert an object into a mapper function
}
