% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{detect}
\alias{detect}
\alias{detect_index}
\title{Find the value or position of the first match}
\usage{
detect(.x, .f, ..., .right = FALSE, .p = is_true)

detect_index(.x, .f, ..., .right = FALSE, .p = is_true)
}
\description{
Find the value or position of the first match
}
