% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every.r
\name{every}
\alias{every}
\alias{some}
\alias{none}
\title{Do every, some, or none of the elements of a list satisfy a predicate?}
\usage{
every(.x, .p, ...)

some(.x, .p, ...)

none(.x, .p, ...)
}
\description{
Do every, some, or none of the elements of a list satisfy a predicate?
}
