% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_flatten.R
\name{flatten_list}
\alias{flatten_list}
\alias{list_flatten}
\title{Flatten a list}
\usage{
list_flatten(
  x,
  ...,
  name_spec = "{outer}_{inner}",
  name_repair = c("minimal", "unique", "check_unique", "universal")
)
}
\description{
Flatten a list
}
