% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_while.R
\name{head_while}
\alias{head_while}
\alias{tail_while}
\title{Find head/tail that all satisfies a predicate.}
\usage{
head_while(.x, .p, ...)

tail_while(.x, .p, ...)
}
\description{
Find head/tail that all satisfies a predicate.
}
