% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imap.R
\name{imap}
\alias{imap}
\alias{iwalk}
\alias{imap_lgl}
\alias{imap_int}
\alias{imap_dbl}
\alias{imap_chr}
\title{Standalone \code{imap} - \code{purrr}-like functions}
\usage{
imap(.x, .f, ...)

iwalk(.x, .f, ...)

imap_lgl(.x, .y, .f, ...)

imap_int(.x, .y, .f, ...)

imap_dbl(.x, .y, .f, ...)

imap_chr(.x, .y, .f, ...)
}
\description{
Standalone \code{imap} - \code{purrr}-like functions
}
