% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.r
\name{keep}
\alias{keep}
\alias{discard}
\alias{compact}
\title{Keep/discard elements based on their values}
\usage{
keep(.x, .f, ...)

discard(.x, .p, ...)

compact(.x)
}
\description{
Keep/discard elements based on their values
}
