% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_at.R
\name{keep_at}
\alias{keep_at}
\alias{discard_at}
\title{Keep/discard elements based on their name/position}
\usage{
keep_at(x, at)

discard_at(x, at)
}
\description{
Keep/discard elements based on their name/position
}
