% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_c.r
\name{list_c}
\alias{list_c}
\alias{list_rbind}
\alias{list_cbind}
\title{Combine list elements into a single data structure}
\usage{
list_c(x)

list_rbind(x)

list_cbind(x)
}
\description{
Combine list elements into a single data structure
}
