% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_simplify.R
\name{list_simplify}
\alias{list_simplify}
\title{Simplify a list to an atomic or S3 vector}
\usage{
list_simplify(
  x,
  ...,
  strict = TRUE,
  error_arg = caller_arg(x),
  error_call = caller_env()
)
}
\description{
Simplify a list to an atomic or S3 vector
}
