% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{walk}
\alias{map_lgl}
\alias{map_int}
\alias{map_dbl}
\alias{map_chr}
\alias{map_raw}
\title{Standalone \code{map} - \code{purrr}-like functions}
\usage{
map(.x, .f, ...)

walk(.x, .f, ...)

map_lgl(.x, .f, ...)

map_int(.x, .f, ...)

map_dbl(.x, .f, ...)

map_chr(.x, .f, ...)

map_raw(.x, .f, ...)
}
\description{
Standalone \code{map} - \code{purrr}-like functions
}
