% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_if.r
\name{map_if}
\alias{map_if}
\title{Apply a function to each element of a vector conditionally}
\usage{
map_if(.x, .p, .f, ..., .else = NULL)
}
\description{
Apply a function to each element of a vector conditionally
}
