% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negate.R
\name{negate}
\alias{negate}
\title{Negate a predicate function so it selects what it previously rejected}
\usage{
negate(.p)
}
\description{
Negate a predicate function so it selects what it previously rejected
}
