% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck.R
\name{pluck}
\alias{pluck}
\alias{pluck<-}
\alias{pluck_exists}
\alias{pluck_depth}
\alias{ith}
\alias{ith_index}
\title{Safely get or set an element deep within a nested data structure}
\usage{
pluck(.x, ...) <- value

pluck(.x, ..., .default = NULL)

pluck_exists(.x, ...)

pluck_depth(x, is_node = NULL)

ith(x, i, default = NULL, arg = caller_arg(i), call = caller_env())

ith_index(i, n, .default = NULL)
}
\description{
Safely get or set an element deep within a nested data structure
}
