% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap.r
\name{pmap}
\alias{pmap}
\alias{pwalk}
\alias{pmap_lgl}
\alias{pmap_int}
\alias{pmap_dbl}
\alias{pmap_chr}
\title{Map over multiple input simultaneously (in "parallel")}
\usage{
pmap(.l, .f, ...)

pwalk(.l, .f, ...)

pmap_lgl(.l, .f, ...)

pmap_int(.l, .f, ...)

pmap_dbl(.l, .f, ...)

pmap_chr(.l, .f, ...)
}
\description{
Map over multiple input simultaneously (in "parallel")
}
