% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poorrr-package.R
\docType{package}
\name{poorrr-package}
\alias{poorrr}
\alias{poorrr-package}
\title{poorrr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/poorrr}
  \item Report bugs at \url{https://github.com/f-dallolio/poorrr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
