% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr.R
\name{purrr-like}
\alias{purrr-like}
\alias{pmap}
\alias{pwalk}
\alias{pmap_lgl}
\alias{pmap_int}
\alias{pmap_dbl}
\alias{pmap_chr}
\alias{keep}
\alias{discard}
\alias{map_if}
\alias{where_if}
\alias{map_at}
\alias{where_at}
\alias{compact}
\alias{list_transpose}
\alias{every}
\alias{some}
\alias{none}
\alias{negate}
\alias{reduce}
\alias{reduce_right}
\alias{accumulate}
\alias{accumulate_right}
\alias{detect}
\alias{detect_index}
\alias{list_c}
\title{Standalone \code{purrr}-like functions}
\usage{
pmap(.l, .f, ...)

pwalk(.l, .f, ...)

pmap_lgl(.l, .f, ...)

pmap_int(.l, .f, ...)

pmap_dbl(.l, .f, ...)

pmap_chr(.l, .f, ...)

keep(.x, .f, ...)

discard(.x, .p, ...)

map_if(.x, .p, .f, ..., .else = NULL)

where_if(.x, .p, ...)

map_at(.x, .at, .f, ...)

where_at(.x, .at, ...)

compact(.x)

list_transpose(.l)

every(.x, .p, ...)

some(.x, .p, ...)

none(.x, .p, ...)

negate(.p)

reduce(.x, .f, ..., .init)

reduce_right(.x, .f, ..., .init)

accumulate(.x, .f, ..., .init)

accumulate_right(.x, .f, ..., .init)

detect(.x, .f, ..., .right = FALSE, .p = is_true)

detect_index(.x, .f, ..., .right = FALSE, .p = is_true)

list_c(x)
}
\description{
Standalone \code{purrr}-like functions
}
