% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.r
\name{reduce}
\alias{reduce}
\title{Reduce a list to a single value by iteratively applying a binary function}
\usage{
reduce(.x, .f, ..., .init)
}
\description{
Reduce a list to a single value by iteratively applying a binary function
}
