% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.R
\name{utils-assn}
\alias{utils-assn}
\alias{get_assn_obj}
\alias{get_assn_name}
\alias{eval_assn_call}
\alias{deparse_assn_obj}
\alias{deparse_assn_call}
\alias{string_assn_call}
\title{Tools for dealing with assignment calls (usually from parse)}
\usage{
get_assn_obj(x, env = globalenv())

get_assn_name(x)

eval_assn_call(x, env = globalenv())

deparse_assn_obj(x, env = globalenv())

deparse_assn_call(x, as_string = FALSE, env = globalenv())

string_assn_call(x, ..., as_string = TRUE, env = globalenv())
}
\description{
Tools for dealing with assignment calls (usually from parse)
}
