% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-predicates.R
\name{utils-bare-predicates}
\alias{utils-bare-predicates}
\alias{is_bare_atomic}
\alias{is_bare_character}
\alias{is_bare_double}
\alias{is_bare_integer}
\alias{is_bare_list}
\alias{is_bare_logical}
\alias{is_bare_numeric}
\alias{is_bare_vector}
\title{Utilities: Type predicates (bare)}
\usage{
is_bare_atomic(x, n = NULL)

is_bare_character(x, n = NULL)

is_bare_double(x, n = NULL)

is_bare_integer(x, n = NULL)

is_bare_list(x, n = NULL)

is_bare_logical(x, n = NULL)

is_bare_numeric(x, n = NULL)

is_bare_vector(x, n = NULL)
}
\description{
Utilities: Type predicates (bare)
}
