% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{utils-env}
\alias{utils-env}
\alias{env_parent}
\alias{env_get_list}
\alias{env_get}
\alias{env_names}
\alias{env_names_if}
\title{Get properties of the current or caller frame}
\usage{
env_parent(env = parent.frame(), n = 1)

env_get_list(
  env = caller_env(),
  nms = names(env),
  default = list(NULL),
  inherit = FALSE,
  last = NULL,
  mode = "any"
)

env_get(
  env = caller_env(),
  nm,
  default = NULL,
  inherit = FALSE,
  last = NULL,
  mode = "any"
)

env_names(env, sort = FALSE)

env_names_if(env, .p = NULL, ..., sort = FALSE)
}
\description{
Get properties of the current or caller frame
}
