% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-frame.R
\name{utils-frame}
\alias{utils-frame}
\alias{current_env}
\alias{caller_env}
\alias{caller_arg}
\title{Get properties of the current or caller frame}
\usage{
current_env()

caller_env(n = 1)

caller_arg(arg)
}
\description{
Get properties of the current or caller frame
}
