% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{utils-ns}
\alias{utils-ns}
\alias{ns_env}
\alias{ns_imports}
\alias{ns_imports_env}
\alias{ns_import_from}
\alias{ns_env_name}
\alias{ns_exports}
\alias{ns_privates}
\alias{ns_names}
\alias{ns_exports_has}
\title{Get the namespace of a package}
\usage{
ns_env(x = parent.frame())

ns_imports(ns, sort = TRUE)

ns_imports_env(x = parent.frame())

ns_import_from(ns, names, env = parent.frame())

ns_env_name(x = parent.frame())

ns_exports(ns, sort = TRUE)

ns_privates(ns, sort = TRUE)

ns_names(ns, sort = TRUE)

ns_exports_has(ns, name)
}
\description{
Get the namespace of a package
}
