% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-predicates.R
\name{utils-scalar-predicates}
\alias{utils-scalar-predicates}
\alias{is_scalar_numeric}
\alias{is_scalar_list}
\alias{is_scalar_logical}
\alias{is_scalar_vector}
\alias{is_string}
\alias{is_bool}
\title{Utilities: Type predicates (scalar)}
\usage{
is_scalar_numeric(x)

is_scalar_list(x)

is_scalar_logical(x)

is_scalar_vector(x)

is_string(x, string = NULL)

is_bool(x)
}
\description{
Utilities: Type predicates (scalar)
}
