% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-standalone.R
\name{utils-standalone}
\alias{utils-standalone}
\alias{standalone_urls}
\alias{standalone_choices}
\alias{standalone_select}
\alias{standalone_get}
\alias{choose_standalones}
\alias{local_standalones}
\alias{standalone_info}
\alias{print.standalone_info}
\title{Tools for managing standalone files from Github.}
\usage{
standalone_urls(repo, sort = TRUE)

standalone_choices(repo, sprt = TRUE)

standalone_select(nm, repo, use_grep = TRUE, sort = TRUE, simplify = TRUE)

standalone_get(
  repo,
  nm = ".",
  dest_dir = "code/standalone",
  ...,
  use_grep = TRUE,
  sort = TRUE,
  simplify = TRUE
)

choose_standalones(
  repo,
  dest_dir = "R",
  multiple = TRUE,
  ...,
  use_grep = TRUE,
  sort = TRUE,
  simplify = TRUE
)

local_standalones(dir, pattern = "standalone-", full_path = FALSE)

standalone_info(
  repo,
  nm = NULL,
  use_grep = TRUE,
  sort = FALSE,
  simplify = TRUE
)

\method{print}{standalone_info}(x)
}
\description{
Tools for managing standalone files from Github.
}
