% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{utils-string}
\alias{utils-string}
\alias{str_flatten_comma}
\alias{str_flatten}
\alias{str_detect}
\alias{str_which}
\alias{str_subset}
\alias{str_replace}
\alias{str_replace_all}
\alias{str_remove}
\alias{str_remove_all}
\alias{str_pad}
\alias{str_numpad}
\alias{str_enum}
\alias{str_enum_left}
\alias{str_enum_right}
\alias{str_split}
\alias{str_split_fixed}
\alias{str_split_1}
\title{Base replacement for string manipulation based on \code{stringr}.}
\usage{
str_flatten_comma(string, last = "and", na.rm = FALSE, oxford = TRUE)

str_flatten(string, collapse = "", ...)

str_detect(
  string,
  pattern,
  negate = FALSE,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_which(
  string,
  pattern,
  negate = FALSE,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_subset(
  string,
  pattern,
  negate = FALSE,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_replace(
  string,
  pattern,
  replacement,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_replace_all(
  string,
  pattern,
  replacement,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_remove(
  string,
  pattern,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_remove_all(
  string,
  pattern,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)

str_pad(string, side = c("left", "right", "both"), pad = " ", use_width = TRUE)

str_numpad(x)

str_enum(x, side = c("right", "left"), sep = "")

str_enum_left(x, sep = "")

str_enum_right(x, sep = "")

str_split(string, pattern, n = Inf, simplify = FALSE)

str_split_fixed(string, pattern, n = Inf)

str_split_1(string, pattern)
}
\description{
Base replacement for string manipulation based on \code{stringr}.
}
