% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-predicates.R
\name{utils-type-predicates}
\alias{utils-type-predicates}
\alias{is_atomic}
\alias{is_character}
\alias{is_double}
\alias{is_integer}
\alias{is_numeric}
\alias{is_list}
\alias{is_logical}
\alias{is_vector}
\title{Utilities: Type predicates}
\usage{
is_atomic(x, n = NULL)

is_character(x, n = NULL)

is_double(x, n = NULL, finite = NULL)

is_integer(x, n = NULL)

is_numeric(x, n = NULL)

is_list(x, n = NULL)

is_logical(x, n = NULL)

is_vector(x, n = NULL)
}
\description{
Utilities: Type predicates
}
