% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.R
\name{accumulate}
\alias{accumulate}
\alias{accumulate_right}
\title{Accumulate intermediate results of a vector reduction}
\usage{
accumulate(.x, .f, ..., .init)

accumulate_right(.x, .f, ..., .init)
}
\description{
Accumulate intermediate results of a vector reduction
}
