% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imap.R
\name{imap}
\alias{imap}
\alias{iwalk}
\alias{imap_lgl}
\alias{imap_int}
\alias{imap_dbl}
\alias{imap_chr}
\title{Apply a function to each element of a vector, and its index}
\usage{
imap(.x, .f, ...)

iwalk(.x, .f, ...)

imap_lgl(.x, .f, ...)

imap_int(.x, .f, ...)

imap_dbl(.x, .f, ...)

imap_chr(.x, .f, ...)
}
\description{
Apply a function to each element of a vector, and its index
}
