% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\name{map2Î}
\alias{map2Î}
\alias{map2}
\alias{walk2}
\alias{map2_lgl}
\alias{map2_int}
\alias{map2_dbl}
\alias{map2_chr}
\title{Standalone \code{map2} - \code{purrr}-like functions}
\usage{
map2(.x, .y, .f, ...)

walk2(.x, .y, .f, ...)

map2_lgl(.x, .y, .f, ...)

map2_int(.x, .y, .f, ...)

map2_dbl(.x, .y, .f, ...)

map2_chr(.x, .y, .f, ...)
}
\description{
Standalone \code{map2} - \code{purrr}-like functions
}
