% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_if.R
\name{map_if}
\alias{map_if}
\alias{map_at}
\title{Apply a function to each element of a vector conditionally}
\usage{
map_if(.x, .p, .f, ...)

map_at(.x, .at, .f, ...)
}
\description{
Apply a function to each element of a vector conditionally
}
