% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poorrr-package.R
\docType{package}
\name{poorrr-package}
\alias{poorrr}
\alias{poorrr-package}
\title{poorrr: What the Package Does (One Line, Title Case)}
\description{
A poor man's dependency-free recreation of package 'purrr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/poorrr}
  \item \url{https://f-dallolio.github.io/poorrr/}
  \item Report bugs at \url{https://github.com/f-dallolio/poorrr/issues}
}

}
\author{
\strong{Maintainer}: Filippo Dall'Olio \email{fdallolio@brocku.ca}

}
\keyword{internal}
