% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poorrr-package.R
\docType{package}
\name{poorrr-package}
\alias{poorrr}
\alias{poorrr-package}
\title{poorrr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
