% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{utils-frame}
\alias{utils-frame}
\alias{current_env}
\alias{caller_env}
\title{Get properties of the current or caller frame}
\usage{
current_env()

caller_env(n = 1)
}
\description{
Get properties of the current or caller frame
}
